+++ src/payments.ios.ts
@@ -5,14 +5,14 @@ import {
   _payments$,
   EventContext,
   EventResult,
-  Event,
+  PaymentEvent,
 } from './payments.common';
 
 export {
   EventContext,
   EventResult,
   IPaymentEvent,
-  Event,
+  PaymentEvent,
   payments$,
 } from './payments.common';
 
@@ -23,31 +23,31 @@ let _paymentTransactionObserver: SKPaymentTransactionObserverImpl | null;
 export function init(): void {
   if (!_paymentTransactionObserver) {
     _payments$.next({
-      context: Event.Context.CONNECTING_STORE,
-      result: Event.Result.STARTED,
+      context: PaymentEvent.Context.CONNECTING_STORE,
+      result: PaymentEvent.Result.STARTED,
       payload: null,
     });
     /* tslint:disable: no-use-before-declare */
     _paymentTransactionObserver = new SKPaymentTransactionObserverImpl();
     /* tslint:enable: no-use-before-declare */
     _payments$.next({
-      context: Event.Context.CONNECTING_STORE,
-      result: Event.Result.PENDING,
+      context: PaymentEvent.Context.CONNECTING_STORE,
+      result: PaymentEvent.Result.PENDING,
       payload: null,
     });
     try {
       SKPaymentQueue.defaultQueue().addTransactionObserver(_paymentTransactionObserver);
       _payments$.next({
-        context: Event.Context.CONNECTING_STORE,
-        result: Event.Result.SUCCESS,
+        context: PaymentEvent.Context.CONNECTING_STORE,
+        result: PaymentEvent.Result.SUCCESS,
         payload: null,
       });
     } catch (e) {
       const errorPayload = typeof e === 'object' ? e.message : e;
       console.error(new Error(`Init failed: ${errorPayload}`));
       _payments$.next({
-        context: Event.Context.CONNECTING_STORE,
-        result: Event.Result.FAILURE,
+        context: PaymentEvent.Context.CONNECTING_STORE,
+        result: PaymentEvent.Result.FAILURE,
         payload: new Failure(null),
       });
     }
@@ -63,8 +63,8 @@ export function tearDown(): void {
 
 export function fetchItems(itemIds: Array<string>): void {
   _payments$.next({
-    context: Event.Context.RETRIEVING_ITEMS,
-    result: Event.Result.STARTED,
+    context: PaymentEvent.Context.RETRIEVING_ITEMS,
+    result: PaymentEvent.Result.STARTED,
     payload: itemIds,
   });
   const productIds: NSMutableSet<string> = NSMutableSet.alloc<string>().init();
@@ -76,8 +76,8 @@ export function fetchItems(itemIds: Array<string>): void {
   _productRequest.delegate = _productRequestDelegate;
   _productRequest.start();
   _payments$.next({
-    context: Event.Context.RETRIEVING_ITEMS,
-    result: Event.Result.PENDING,
+    context: PaymentEvent.Context.RETRIEVING_ITEMS,
+    result: PaymentEvent.Result.PENDING,
     payload: itemIds,
   });
 }
@@ -90,8 +90,8 @@ export function buyItem(
     const pendingCount = SKPaymentQueue.defaultQueue().transactions.count;
     if (!pendingCount) {
       _payments$.next({
-        context: Event.Context.PROCESSING_ORDER,
-        result: Event.Result.PENDING,
+        context: PaymentEvent.Context.PROCESSING_ORDER,
+        result: PaymentEvent.Result.PENDING,
         payload: pendingCount + 1,
       });
       const payment = SKMutablePayment.paymentWithProduct(<SKProduct>item.nativeValue);
@@ -101,23 +101,23 @@ export function buyItem(
       try {
         SKPaymentQueue.defaultQueue().addPayment(payment);
         _payments$.next({
-          context: Event.Context.PROCESSING_ORDER,
-          result: Event.Result.STARTED,
+          context: PaymentEvent.Context.PROCESSING_ORDER,
+          result: PaymentEvent.Result.STARTED,
           payload: item,
         });
       } catch (e) {
         const errorPayload = typeof e === 'object' ? e.message : e;
         console.error(new Error(`Error while adding payment: ${errorPayload}`));
         _payments$.next({
-          context: Event.Context.PROCESSING_ORDER,
-          result: Event.Result.FAILURE,
+          context: PaymentEvent.Context.PROCESSING_ORDER,
+          result: PaymentEvent.Result.FAILURE,
           payload: new Failure(null),
         });
       }
     } else {
       _payments$.next({
-        context: Event.Context.PROCESSING_ORDER,
-        result: Event.Result.PENDING,
+        context: PaymentEvent.Context.PROCESSING_ORDER,
+        result: PaymentEvent.Result.PENDING,
         payload: pendingCount,
       });
     }
@@ -128,31 +128,31 @@ export function buyItem(
 
 export function finalizeOrder(order: Order): void {
   _payments$.next({
-    context: Event.Context.FINALIZING_ORDER,
-    result: Event.Result.STARTED,
+    context: PaymentEvent.Context.FINALIZING_ORDER,
+    result: PaymentEvent.Result.STARTED,
     payload: order,
   });
   if (order.state === OrderState.VALID && !order.restored) {
     try {
       SKPaymentQueue.defaultQueue().finishTransaction(<SKPaymentTransaction>order.nativeValue);
       _payments$.next({
-        context: Event.Context.FINALIZING_ORDER,
-        result: Event.Result.PENDING,
+        context: PaymentEvent.Context.FINALIZING_ORDER,
+        result: PaymentEvent.Result.PENDING,
         payload: order,
       });
     } catch (e) {
       const errorPayload = typeof e === 'object' ? e.message : e;
       console.error(new Error(`Error while finalizing order: ${errorPayload}`));
       _payments$.next({
-        context: Event.Context.FINALIZING_ORDER,
-        result: Event.Result.FAILURE,
+        context: PaymentEvent.Context.FINALIZING_ORDER,
+        result: PaymentEvent.Result.FAILURE,
         payload: new Failure(null),
       });
     }
   } else {
     _payments$.next({
-      context: Event.Context.FINALIZING_ORDER,
-      result: Event.Result.FAILURE,
+      context: PaymentEvent.Context.FINALIZING_ORDER,
+      result: PaymentEvent.Result.FAILURE,
       payload: new Failure(999),
     });
   }
@@ -160,8 +160,8 @@ export function finalizeOrder(order: Order): void {
 
 export function restoreOrders(): void {
   _payments$.next({
-    context: Event.Context.RESTORING_ORDERS,
-    result: Event.Result.STARTED,
+    context: PaymentEvent.Context.RESTORING_ORDERS,
+    result: PaymentEvent.Result.STARTED,
     payload: null,
   });
   try {
@@ -170,8 +170,8 @@ export function restoreOrders(): void {
     const errorPayload = typeof e === 'object' ? e.message : e;
     console.error(new Error(`Error while restoring order: ${errorPayload}`));
     _payments$.next({
-      context: Event.Context.RESTORING_ORDERS,
-      result: Event.Result.FAILURE,
+      context: PaymentEvent.Context.RESTORING_ORDERS,
+      result: PaymentEvent.Result.FAILURE,
       payload: new Failure(null),
     });
   }
@@ -201,8 +201,8 @@ class SKProductRequestDelegateImpl extends NSObject implements SKProductsRequest
     }
 
     _payments$.next({
-      context: Event.Context.RETRIEVING_ITEMS,
-      result: Event.Result.SUCCESS,
+      context: PaymentEvent.Context.RETRIEVING_ITEMS,
+      result: PaymentEvent.Result.SUCCESS,
       payload: result,
     });
     this._cleanup();
@@ -213,8 +213,8 @@ class SKProductRequestDelegateImpl extends NSObject implements SKProductsRequest
     error: NSError,
   ) {
     _payments$.next({
-      context: Event.Context.RETRIEVING_ITEMS,
-      result: Event.Result.FAILURE,
+      context: PaymentEvent.Context.RETRIEVING_ITEMS,
+      result: PaymentEvent.Result.FAILURE,
       payload: new Failure(error.code),
     });
     this._cleanup();
@@ -241,8 +241,8 @@ class SKPaymentTransactionObserverImpl extends NSObject implements SKPaymentTran
 
   public paymentQueueRestoreCompletedTransactionsFinished(queue: SKPaymentQueue): void {
     _payments$.next({
-      context: Event.Context.RESTORING_ORDERS,
-      result: Event.Result.SUCCESS,
+      context: PaymentEvent.Context.RESTORING_ORDERS,
+      result: PaymentEvent.Result.SUCCESS,
       payload: null,
     });
   }
@@ -252,8 +252,8 @@ class SKPaymentTransactionObserverImpl extends NSObject implements SKPaymentTran
     error: NSError,
   ): void {
     _payments$.next({
-      context: Event.Context.RESTORING_ORDERS,
-      result: Event.Result.FAILURE,
+      context: PaymentEvent.Context.RESTORING_ORDERS,
+      result: PaymentEvent.Result.FAILURE,
       payload: new Failure(error.code),
     });
   }
@@ -267,16 +267,16 @@ class SKPaymentTransactionObserverImpl extends NSObject implements SKPaymentTran
         const transaction: SKPaymentTransaction = transactions.objectAtIndex(i);
         if (transaction.transactionState === SKPaymentTransactionState.Purchased) {
           _payments$.next({
-            context: Event.Context.FINALIZING_ORDER,
-            result: Event.Result.SUCCESS,
+            context: PaymentEvent.Context.FINALIZING_ORDER,
+            result: PaymentEvent.Result.SUCCESS,
             payload: new Order(transaction),
           });
         }
       }
     }
     _payments$.next({
-      context: Event.Context.PROCESSING_ORDER,
-      result: Event.Result.PENDING,
+      context: PaymentEvent.Context.PROCESSING_ORDER,
+      result: PaymentEvent.Result.PENDING,
       payload: queue.transactions ? queue.transactions.count : 0,
     });
   }
@@ -302,8 +302,8 @@ function _transactionHandler(
   transactions: NSArray<SKPaymentTransaction>,
 ): void {
   _payments$.next({
-    context: Event.Context.PROCESSING_ORDER,
-    result: Event.Result.PENDING,
+    context: PaymentEvent.Context.PROCESSING_ORDER,
+    result: PaymentEvent.Result.PENDING,
     payload: queue.transactions ? queue.transactions.count : 0,
   });
   if (transactions && transactions.count) {
@@ -313,15 +313,15 @@ function _transactionHandler(
       switch (transaction.transactionState) {
         case SKPaymentTransactionState.Purchased:
           _payments$.next({
-            context: Event.Context.PROCESSING_ORDER,
-            result: Event.Result.SUCCESS,
+            context: PaymentEvent.Context.PROCESSING_ORDER,
+            result: PaymentEvent.Result.SUCCESS,
             payload: new Order(transaction),
           });
           break;
         case SKPaymentTransactionState.Failed:
           _payments$.next({
-            context: Event.Context.PROCESSING_ORDER,
-            result: Event.Result.FAILURE,
+            context: PaymentEvent.Context.PROCESSING_ORDER,
+            result: PaymentEvent.Result.FAILURE,
             payload: new Failure(transaction.error.code),
           });
           try {
@@ -333,13 +333,13 @@ function _transactionHandler(
           break;
         case SKPaymentTransactionState.Restored:
           _payments$.next({
-            context: Event.Context.PROCESSING_ORDER,
-            result: Event.Result.SUCCESS,
+            context: PaymentEvent.Context.PROCESSING_ORDER,
+            result: PaymentEvent.Result.SUCCESS,
             payload: new Order(transaction.originalTransaction, true),
           });
           _payments$.next({
-            context: Event.Context.RESTORING_ORDERS,
-            result: Event.Result.PENDING,
+            context: PaymentEvent.Context.RESTORING_ORDERS,
+            result: PaymentEvent.Result.PENDING,
             payload: new Order(transaction.originalTransaction, true),
           });
           try {
@@ -359,8 +359,8 @@ function _transactionHandler(
     }
   }
   _payments$.next({
-    context: Event.Context.PROCESSING_ORDER,
-    result: Event.Result.PENDING,
+    context: PaymentEvent.Context.PROCESSING_ORDER,
+    result: PaymentEvent.Result.PENDING,
     payload: queue.transactions ? queue.transactions.count : 0,
   });
 }